Refactor waterfall visualization
